chore: redirect all new free analysis to /analysis endpoint
