+++ src/components/Analysis/index.ts
@@ -12,4 +12,3 @@ export * from './PositionEvaluationContainer'
 export * from './VerticalEvaluationBar'
 export * from './MoveRecommendations'
 export * from './MovesByRating'
-export * from './ClientAnalysisGameList'
