+++ src/components/Board/index.ts
@@ -4,7 +4,6 @@ export * from './MovesContainer'
 export * from './BoardController'
 export * from './PromotionOverlay'
 export * from './GameplayInterface'
-export * from './ClientGameplayInterface'
-export * from './ClientMovesContainer'
-export * from './ClientGameBoard'
-export * from './ClientBoardController'
+export * from './AnalysisMovesContainer'
+export * from './AnalysisGameBoard'
+export * from './AnalysisBoardController'
