--- src/contexts/ClientGameControllerContext/index.ts
@@ -1,1 +1,0 @@
-export * from './ClientGameControllerContext'
