+++ src/contexts/index.ts
@@ -5,4 +5,4 @@ export * from './GameControllerContext'
 export * from './AuthContext'
 export * from './TuringControllerContext'
 export * from './AnalysisListContext'
-export * from './ClientGameControllerContext'
+export * from './AnalysisGameControllerContext'
