+++ src/hooks/index.ts
@@ -9,5 +9,4 @@ export * from './useStockfishEngine'
 export * from './useMaiaEngine'
 export * from './useAnalysisController'
 export * from './useLegacyAnalysisController'
-export * from './useClientGameController'
-export * from './useClientAnalysisController'
+export * from './useAnalysisGameController'
