+++ src/hooks/useAnalysisController/useAnalysisController.ts
@@ -1,14 +1,17 @@
 import { Chess } from 'chess.ts'
-import { useEffect, useMemo, useState, useCallback } from 'react'
+import React, { useEffect, useMemo, useState, useCallback } from 'react'
 
 import {
-  Color,
-  MoveMap,
-  MaiaEvaluation,
+  GameTree,
   AnalyzedGame,
+  MaiaEvaluation,
   StockfishEvaluation,
 } from 'src/types'
-import { useGameController, useStockfishEngine, useMaiaEngine } from '..'
+import {
+  useStockfishEngine,
+  useMaiaEngine,
+  useAnalysisGameController,
+} from '..'
 
 const MAIA_MODELS = [
   'maia_kdd_1100',
@@ -25,12 +28,13 @@ const MAIA_MODELS = [
 const MAIA_COLORS = ['#fe7f6d', '#f08a4c', '#ecaa4f', '#eccd4f']
 const STOCKFISH_COLORS = ['#A3C6F8', '#8fadd9', '#7a95ba', '#667c9b']
 
-export const useAnalysisController = (
-  game: AnalyzedGame,
-  initialIndex: number,
-  initialOrientation: Color,
-) => {
-  const controller = useGameController(game, initialIndex, initialOrientation)
+export const useAnalysisController = (game: AnalyzedGame) => {
+  const controller = useAnalysisGameController(
+    game.tree as GameTree,
+    game.tree?.getRoot(),
+  )
+
+  const [analysisState, setAnalysisState] = useState(0)
 
   const {
     maia,
@@ -39,34 +43,21 @@ export const useAnalysisController = (
     progress: maiaProgress,
     downloadModel: downloadMaia,
   } = useMaiaEngine()
+
   const { streamEvaluations, stopEvaluation } = useStockfishEngine()
   const [currentMove, setCurrentMove] = useState<[string, string] | null>()
-  const [stockfishEvaluations, setStockfishEvaluations] = useState<
-    StockfishEvaluation[]
-  >([])
-  const [maiaEvaluations, setMaiaEvaluations] = useState<
-    { [rating: string]: MaiaEvaluation }[]
-  >([])
   const [currentMaiaModel, setCurrentMaiaModel] = useState(MAIA_MODELS[0])
 
-  const parseStockfishEvaluation = useCallback(
-    (evaluation: StockfishEvaluation, fen: string) => {
-      setStockfishEvaluations((prev) => {
-        const newEvaluations = [...prev]
-        const moveIndex = game.moves.findIndex((move) => move.board === fen)
-        newEvaluations[moveIndex] = evaluation
-
-        return newEvaluations
-      })
-    },
-    [game.moves],
-  )
-
   useEffect(() => {
-    const board = new Chess(game.moves[controller.currentIndex].board)
+    if (!controller.currentNode) return
 
+    const board = new Chess(controller.currentNode.fen)
     ;(async () => {
-      if (maiaStatus !== 'ready' || maiaEvaluations[controller.currentIndex])
+      if (
+        maiaStatus !== 'ready' ||
+        !controller.currentNode ||
+        controller.currentNode.analysis.maia
+      )
         return
 
       const { result } = await maia.batchEvaluate(
@@ -75,32 +66,24 @@ export const useAnalysisController = (
         [1100, 1200, 1300, 1400, 1500, 1600, 1700, 1800, 1900],
       )
 
-      const output = {
-        maia_kdd_1100: result[0],
-        maia_kdd_1200: result[1],
-        maia_kdd_1300: result[2],
-        maia_kdd_1400: result[3],
-        maia_kdd_1500: result[4],
-        maia_kdd_1600: result[5],
-        maia_kdd_1700: result[6],
-        maia_kdd_1800: result[7],
-        maia_kdd_1900: result[8],
-      }
-
-      setMaiaEvaluations((prev) => {
-        const newEvaluations = [...prev]
-        newEvaluations[controller.currentIndex] = output
-
-        return newEvaluations
+      const maiaEval: { [key: string]: MaiaEvaluation } = {}
+      MAIA_MODELS.forEach((model, index) => {
+        maiaEval[model] = result[index]
       })
+
+      controller.currentNode.addMaiaAnalysis(maiaEval)
+      setAnalysisState((state) => state + 1)
     })()
-  }, [controller.currentIndex, game.type, status])
+  }, [maiaStatus, controller.currentNode, analysisState])
 
   useEffect(() => {
-    if (game.type === 'tournament') return
-
-    const board = new Chess(game.moves[controller.currentIndex].board)
-    if (stockfishEvaluations[controller.currentIndex]?.depth == 18) return
+    if (!controller.currentNode) return
+    const board = new Chess(controller.currentNode.fen)
+    if (
+      controller.currentNode.analysis.stockfish &&
+      controller.currentNode.analysis.stockfish?.depth >= 18
+    )
+      return
 
     const evaluationStream = streamEvaluations(
       board.fen(),
@@ -110,7 +93,12 @@ export const useAnalysisController = (
     if (evaluationStream) {
       ;(async () => {
         for await (const evaluation of evaluationStream) {
-          parseStockfishEvaluation(evaluation, board.fen())
+          if (!controller.currentNode) {
+            stopEvaluation()
+            break
+          }
+          controller.currentNode.addStockfishAnalysis(evaluation)
+          setAnalysisState((state) => state + 1)
         }
       })()
     }
@@ -118,21 +106,14 @@ export const useAnalysisController = (
     return () => {
       stopEvaluation()
     }
-  }, [
-    controller.currentIndex,
-    game.moves,
-    game.type,
-    streamEvaluations,
-    stopEvaluation,
-    parseStockfishEvaluation,
-  ])
+  }, [controller.currentNode, game.type, streamEvaluations, stopEvaluation])
 
   const moves = useMemo(() => {
-    const moveMap = new Map<string, string[]>()
+    if (!controller.currentNode) return new Map<string, string[]>()
 
-    const moves = new Chess(game.moves[controller.currentIndex].board).moves({
-      verbose: true,
-    })
+    const moveMap = new Map<string, string[]>()
+    const chess = new Chess(controller.currentNode.fen)
+    const moves = chess.moves({ verbose: true })
 
     moves.forEach((key) => {
       const { from, to } = key
@@ -140,24 +121,26 @@ export const useAnalysisController = (
     })
 
     return moveMap
-  }, [controller.currentIndex, game.maiaEvaluations, MAIA_MODELS])
+  }, [controller.currentNode])
 
   const colorSanMapping = useMemo(() => {
+    if (!controller.currentNode) return {}
+
     const mapping: {
       [move: string]: {
         san: string
         color: string
       }
     } = {}
-    const moves = new Chess(game.moves[controller.currentIndex].board).moves({
-      verbose: true,
-    })
+
+    const chess = new Chess(controller.currentNode.fen)
+    const moves = chess.moves({ verbose: true })
 
     for (const move of moves) {
       let color, maiaRank, stockfishRank
       const lan = move.from + move.to + (move.promotion || '')
-      const maia = maiaEvaluations[controller.currentIndex]
-      const stockfish = stockfishEvaluations[controller.currentIndex]
+      const maia = controller.currentNode.analysis.maia?.[currentMaiaModel]
+      const stockfish = controller.currentNode.analysis.stockfish
 
       if (stockfish) {
         stockfishRank = Object.keys(stockfish.cp_vec).indexOf(lan) + 1
@@ -168,10 +151,7 @@ export const useAnalysisController = (
       }
 
       if (maia) {
-        maiaRank =
-          Object.keys(
-            maiaEvaluations[controller.currentIndex][currentMaiaModel].policy,
-          ).indexOf(lan) + 1
+        maiaRank = Object.keys(maia.policy).indexOf(lan) + 1
 
         if (maiaRank <= 4) {
           color = MAIA_COLORS[maiaRank]
@@ -185,100 +165,39 @@ export const useAnalysisController = (
     }
 
     return mapping
-  }, [
-    maiaEvaluations,
-    stockfishEvaluations,
-    controller.currentIndex,
-    game.moves,
-  ])
+  }, [currentMaiaModel, controller.currentNode, analysisState])
 
   const moveEvaluation = useMemo(() => {
-    let stockfish
+    if (!controller.currentNode) return null
 
-    if (game.type !== 'tournament') {
-      stockfish = stockfishEvaluations[controller.currentIndex]
-    } else {
-      const lastMove = game.moves[controller.currentIndex].lastMove?.join('')
-      if (!lastMove) return null
-
-      const cp_vec = Object.fromEntries(
-        Object.entries(
-          game.stockfishEvaluations[controller.currentIndex] as MoveMap,
-        ).sort(([, a], [, b]) => b - a),
-      )
-
-      if (cp_vec) {
-        const model_optimal_cp = Math.max(...Object.values(cp_vec))
-
-        const cp_relative_vec = Object.fromEntries(
-          Object.entries(cp_vec).map(([move, evaluation]) => [
-            move,
-            model_optimal_cp - evaluation,
-          ]),
-        )
-
-        stockfish = {
-          cp_vec,
-          cp_relative_vec,
-          model_optimal_cp,
-        } as StockfishEvaluation
-      }
-    }
-
-    let maia
-    if (maiaEvaluations[controller.currentIndex]) {
-      maia = maiaEvaluations[controller.currentIndex][currentMaiaModel]
-    }
+    const stockfish = controller.currentNode.analysis
+      .stockfish as StockfishEvaluation
+    const maia = controller.currentNode.analysis.maia?.[
+      currentMaiaModel
+    ] as MaiaEvaluation
 
     return {
       maia,
       stockfish,
     }
-  }, [
-    game.id,
-    game.moves,
-    maiaEvaluations,
-    currentMaiaModel,
-    stockfishEvaluations,
-    game.stockfishEvaluations,
-    controller.currentIndex,
-  ])
+  }, [currentMaiaModel, controller.currentNode, analysisState])
 
   const blunderMeter = useMemo(() => {
-    let blunderMoveChance = 0
-    let okMoveChance = 0
-    let goodMoveChance = 0
     if (!moveEvaluation || !moveEvaluation.maia || !moveEvaluation.stockfish) {
       return { blunderMoveChance: 0, okMoveChance: 0, goodMoveChance: 0 }
     }
 
+    let blunderMoveChance = 0
+    let okMoveChance = 0
+    let goodMoveChance = 0
+
     const { maia, stockfish } = moveEvaluation
 
     if (game.type === 'tournament') {
-      const max = Math.max(...Object.values(stockfish.cp_vec))
-      const sf = Object.fromEntries(
-        Object.entries(stockfish.cp_vec).map(([key, value]) => [
-          key,
-          max - value,
-        ]),
-      )
-
-      for (const [move, prob] of Object.entries(maia.policy)) {
-        const loss = sf[move]
-
-        if (loss === undefined) continue
-
-        if (loss <= 50) {
-          goodMoveChance += prob * 100
-        } else if (loss <= 150) {
-          okMoveChance += prob * 100
-        } else {
-          blunderMoveChance += prob * 100
-        }
-      }
-
       return { blunderMoveChance, okMoveChance, goodMoveChance }
     } else {
+      if (!maia || !stockfish)
+        return { blunderMoveChance, okMoveChance, goodMoveChance }
       for (const [move, prob] of Object.entries(maia.policy)) {
         const loss = stockfish.cp_relative_vec[move]
 
@@ -298,12 +217,13 @@ export const useAnalysisController = (
   }, [moveEvaluation])
 
   const movesByRating = useMemo(() => {
-    if (!maiaEvaluations[controller.currentIndex]) return
-
-    const maia = maiaEvaluations[controller.currentIndex]
+    if (!controller.currentNode) return
+    const maia = controller.currentNode.analysis.maia
     const stockfish = moveEvaluation?.stockfish
 
     const candidates: string[][] = []
+
+    if (!maia) return
 
     // Get top 3 Maia moves from selected rating level
     for (const move of Object.keys(maia[currentMaiaModel].policy).slice(0, 3)) {
@@ -341,15 +261,10 @@ export const useAnalysisController = (
     }
 
     return data
-  }, [
-    controller.currentIndex,
-    maiaEvaluations,
-    stockfishEvaluations,
-    moveEvaluation,
-    currentMaiaModel,
-  ])
+  }, [currentMaiaModel, moveEvaluation, controller.currentNode, analysisState])
 
   const moveRecommendations = useMemo(() => {
+    if (!moveEvaluation) return {}
     const recommendations: {
       maia?: { move: string; prob: number }[]
       stockfish?: { move: string; cp: number }[]
@@ -374,7 +289,7 @@ export const useAnalysisController = (
     }
 
     return recommendations
-  }, [controller.currentIndex, maiaEvaluations, stockfishEvaluations])
+  }, [moveEvaluation])
 
   const moveMap = useMemo(() => {
     if (!moveEvaluation?.maia || !moveEvaluation?.stockfish) {
@@ -406,29 +321,33 @@ export const useAnalysisController = (
     }
 
     return data
-  }, [controller.currentIndex, moveEvaluation])
+  }, [moveEvaluation])
 
   const move = useMemo(() => {
-    if (
-      currentMove &&
-      game.availableMoves[controller.currentIndex][currentMove.join('')]
-    ) {
-      const {
-        board: fen,
-        check,
-        ...rest
-      } = game.availableMoves[controller.currentIndex][currentMove.join('')]
-      return { move: currentMove, fen, check, ...rest }
+    if (!currentMove) return undefined
+
+    const chess = new Chess(controller.currentNode?.fen)
+    const san = chess.move({ from: currentMove[0], to: currentMove[1] })?.san
+
+    if (san) {
+      return {
+        move: currentMove,
+        fen: chess.fen(),
+        check: chess.inCheck(),
+        san,
+      }
     }
-  }, [controller.currentIndex, currentMove, game.availableMoves])
+
+    return undefined
+  }, [currentMove, controller.currentNode])
 
   return {
     maiaStatus,
+    controller,
     downloadMaia,
     maiaProgress,
     move,
     moves,
-    controller,
     currentMaiaModel,
     setCurrentMaiaModel,
     currentMove,
@@ -439,7 +358,5 @@ export const useAnalysisController = (
     moveRecommendations,
     moveMap,
     blunderMeter,
-    stockfishEvaluations,
-    maiaEvaluations,
   }
 }
