--- src/hooks/useClientAnalysisController/index.ts
@@ -1,1 +1,0 @@
-export * from './useClientAnalysisController'
