+++ src/pages/analysis/[...id].tsx
@@ -1,4 +1,3 @@
-import toast from 'react-hot-toast'
 import React, {
   useRef,
   useMemo,
@@ -9,13 +8,6 @@ import React, {
   useCallback,
   SetStateAction,
 } from 'react'
-import Head from 'next/head'
-import type { NextPage } from 'next'
-import { AnimatePresence, motion } from 'framer-motion'
-import { useRouter } from 'next/router'
-import type { Key } from 'chessground/types'
-import type { DrawBrushes, DrawShape } from 'chessground/draw'
-
 import {
   getLichessGamePGN,
   getAnalyzedUserGame,
@@ -23,33 +15,43 @@ import {
   getAnalyzedTournamentGame,
 } from 'src/api'
 import {
+  PlayedGame,
+  AnalyzedGame,
+  MaiaEvaluation,
+  StockfishEvaluation,
+  GameNode,
+} from 'src/types'
+import {
+  ModalContext,
+  WindowSizeContext,
+  AnalysisGameControllerContext,
+} from 'src/contexts'
+import {
   Loading,
   MoveMap,
   GameInfo,
   Highlight,
-  GameBoard,
   ExportGame,
   MovesByRating,
-  MovesContainer,
-  BoardController,
-  LegacyAnalysisGameList,
-  ContinueAgainstMaia,
-  MoveRecommendations,
-  AuthenticatedWrapper,
-  VerticalEvaluationBar,
   AnalysisGameList,
-  HorizontalEvaluationBar,
+  AnalysisGameBoard,
   DownloadModelModal,
+  MoveRecommendations,
+  ContinueAgainstMaia,
+  AuthenticatedWrapper,
+  AnalysisMovesContainer,
+  VerticalEvaluationBar,
+  HorizontalEvaluationBar,
+  AnalysisBoardController,
 } from 'src/components'
-import { Color, PlayedGame } from 'src/types'
+import Head from 'next/head'
+import type { NextPage } from 'next'
+import toast from 'react-hot-toast'
+import { useRouter } from 'next/router'
+import type { Key } from 'chessground/types'
 import { useAnalysisController } from 'src/hooks'
-import {
-  AnalyzedGame,
-  MaiaEvaluation,
-  StockfishEvaluation,
-} from 'src/types/analysis'
-import { ModalContext, WindowSizeContext } from 'src/contexts'
-import { GameControllerContext } from 'src/contexts/GameControllerContext/GameControllerContext'
+import { AnimatePresence, motion } from 'framer-motion'
+import type { DrawBrushes, DrawShape } from 'chessground/draw'
 import { ConfigureAnalysis } from 'src/components/Analysis/ConfigureAnalysis'
 
 const MAIA_MODELS = [
@@ -76,7 +78,7 @@ const AnalysisPage: NextPage = () => {
   }, [setInstructionsModalProps, openedModals.analysis])
 
   const router = useRouter()
-  const { id, index, orientation } = router.query
+  const { id } = router.query
 
   const [analyzedGame, setAnalyzedGame] = useState<AnalyzedGame | undefined>(
     undefined,
@@ -99,7 +101,9 @@ const AnalysisPage: NextPage = () => {
 
       setAnalyzedGame({ ...game, type: 'tournament' })
       setCurrentId(newId)
-      router.push(`/analysis/${newId.join('/')}`, undefined, { shallow: true })
+      router.push(`/analysis/${newId.join('/')}`, undefined, {
+        shallow: true,
+      })
     },
     [router],
   )
@@ -146,7 +150,9 @@ const AnalysisPage: NextPage = () => {
 
       setAnalyzedGame({ ...game, type })
       setCurrentId([id, type])
-      router.push(`/analysis/${id}/${type}`, undefined, { shallow: true })
+      router.push(`/analysis/${id}/${type}`, undefined, {
+        shallow: true,
+      })
     },
     [],
   )
@@ -179,9 +185,6 @@ const AnalysisPage: NextPage = () => {
         <Analysis
           currentId={currentId}
           analyzedGame={analyzedGame}
-          setAnalyzedGame={setAnalyzedGame}
-          initialIndex={index ? Number(index) : 0}
-          initialOrientation={orientation == 'black' ? 'black' : 'white'}
           getAndSetTournamentGame={getAndSetTournamentGame}
           getAndSetLichessGames={getAndSetLichessGames}
           getAndSetUserGames={getAndSetUserGame}
@@ -195,6 +198,8 @@ const AnalysisPage: NextPage = () => {
 
 interface Props {
   currentId: string[]
+  analyzedGame: AnalyzedGame
+
   getAndSetTournamentGame: (
     newId: string[],
     setCurrentMove?: Dispatch<SetStateAction<number>>,
@@ -209,17 +214,11 @@ interface Props {
     type: 'play' | 'hand' | 'brain',
     setCurrentMove?: Dispatch<SetStateAction<number>>,
   ) => Promise<void>
-  analyzedGame: AnalyzedGame
-  initialIndex: number
-  initialOrientation: Color
-  setAnalyzedGame: Dispatch<SetStateAction<AnalyzedGame | undefined>>
 }
 
 const Analysis: React.FC<Props> = ({
   currentId,
   analyzedGame,
-  initialIndex,
-  initialOrientation,
   getAndSetTournamentGame,
   getAndSetLichessGames,
   getAndSetUserGames,
@@ -238,25 +237,6 @@ const Analysis: React.FC<Props> = ({
       name: 'Export',
     },
   ]
-  const {
-    move,
-    moves,
-    controller,
-    setCurrentMaiaModel,
-    currentMaiaModel,
-    moveEvaluation,
-    setCurrentMove,
-    stockfishEvaluations,
-    maiaEvaluations,
-    blunderMeter,
-    moveMap,
-    movesByRating,
-    colorSanMapping,
-    moveRecommendations,
-    maiaStatus,
-    maiaProgress,
-    downloadMaia,
-  } = useAnalysisController(analyzedGame, initialIndex, initialOrientation)
 
   const { width } = useContext(WindowSizeContext)
   const isMobile = useMemo(() => width > 0 && width <= 670, [width])
@@ -265,10 +245,28 @@ const Analysis: React.FC<Props> = ({
   const [brushes, setBrushes] = useState<DrawBrushes>({} as DrawBrushes)
   const [screen, setScreen] = useState(screens[0])
   const toastId = useRef<string>(null)
+  const [currentSquare, setCurrentSquare] = useState<Key | null>(null)
+
+  const {
+    maiaStatus,
+    downloadMaia,
+    maiaProgress,
+    controller,
+    move,
+    moves,
+    currentMaiaModel,
+    setCurrentMaiaModel,
+    colorSanMapping,
+    moveEvaluation,
+    movesByRating,
+    moveRecommendations,
+    moveMap,
+    blunderMeter,
+  } = useAnalysisController(analyzedGame)
 
   useEffect(() => {
     setHoverArrow(null)
-  }, [controller.currentIndex])
+  }, [controller.currentNode])
 
   useEffect(() => {
     if (maiaStatus === 'loading' && !toastId.current) {
@@ -285,11 +283,11 @@ const Analysis: React.FC<Props> = ({
   }, [maiaStatus])
 
   const launchContinue = useCallback(() => {
-    const fen = analyzedGame.moves[controller.currentIndex].board
+    const fen = controller.currentNode?.fen as string
     const url = '/play' + '?fen=' + encodeURIComponent(fen)
 
     window.open(url)
-  }, [analyzedGame.moves, controller])
+  }, [controller.currentNode])
 
   useEffect(() => {
     const arr = []
@@ -318,7 +316,7 @@ const Analysis: React.FC<Props> = ({
     }
 
     setArrows(arr)
-  }, [stockfishEvaluations, maiaEvaluations, controller.currentIndex])
+  }, [moveEvaluation, controller.currentNode, controller.orientation])
 
   const Player = ({
     name,
@@ -382,15 +380,16 @@ const Analysis: React.FC<Props> = ({
               <div className="flex flex-col items-start">
                 <div className="flex flex-row items-start">
                   <div className="relative flex aspect-square w-[60vh]">
-                    <GameBoard
+                    <AnalysisGameBoard
                       game={analyzedGame}
                       moves={moves}
-                      setCurrentMove={setCurrentMove}
-                      move={move}
-                      brushes={brushes}
+                      setCurrentSquare={setCurrentSquare}
                       shapes={
                         hoverArrow ? [...arrows, hoverArrow] : [...arrows]
                       }
+                      currentNode={controller.currentNode as GameNode}
+                      orientation={controller.orientation}
+                      goToNode={controller.goToNode}
                     />
                   </div>
                   <VerticalEvaluationBar
@@ -406,9 +405,8 @@ const Analysis: React.FC<Props> = ({
                       moveEvaluation?.stockfish
                         ? 600 +
                           moveEvaluation.stockfish.model_optimal_cp *
-                            (analyzedGame.moves[
-                              controller.currentIndex
-                            ].board.split(' ')[1] !== controller.orientation[0]
+                            (analyzedGame.moves[0].board.split(' ')[1] !==
+                            controller.orientation[0]
                               ? 1
                               : -1)
                         : void 0
@@ -422,14 +420,14 @@ const Analysis: React.FC<Props> = ({
                 name={
                   controller.orientation === 'white'
                     ? analyzedGame.whitePlayer.name
-                    : analyzedGame.blackPlayer.name
+                    : analyzedGame.whitePlayer.name
                 }
                 rating={
                   controller.orientation === 'white'
                     ? analyzedGame.whitePlayer.rating
-                    : analyzedGame.blackPlayer.rating
+                    : analyzedGame.whitePlayer.rating
                 }
-                color={controller.orientation === 'white' ? 'white' : 'black'}
+                color={controller.orientation === 'white' ? 'black' : 'white'}
                 termination={analyzedGame.termination.winner}
               />
             </div>
@@ -524,12 +522,10 @@ const Analysis: React.FC<Props> = ({
                 colorSanMapping={colorSanMapping}
               />
               <div className="ovrerflow-y-scroll flex h-[30vh] flex-col">
-                <BoardController setCurrentMove={setCurrentMove} />
-                <MovesContainer
+                <AnalysisBoardController />
+                <AnalysisMovesContainer
                   game={analyzedGame}
-                  setCurrentMove={setCurrentMove}
                   termination={analyzedGame.termination}
-                  currentMaiaModel={currentMaiaModel}
                 />
               </div>
             </div>
@@ -547,25 +543,24 @@ const Analysis: React.FC<Props> = ({
             <></>
           </GameInfo>
           <div className="relative flex h-[100vw] w-screen">
-            <GameBoard
+            <AnalysisGameBoard
               game={analyzedGame}
               moves={moves}
-              setCurrentMove={setCurrentMove}
-              move={move}
-              brushes={brushes}
+              setCurrentSquare={setCurrentSquare}
               shapes={hoverArrow ? [...arrows, hoverArrow] : [...arrows]}
+              currentNode={controller.currentNode as GameNode}
+              orientation={controller.orientation}
+              goToNode={controller.goToNode}
             />
           </div>
           <div className="flex h-auto w-full flex-col gap-1">
             <div className="w-screen !flex-grow-0">
-              <BoardController setCurrentMove={setCurrentMove} />
+              <AnalysisBoardController />
             </div>
             <div className="relative bottom-0 h-full flex-1 overflow-auto">
-              <MovesContainer
+              <AnalysisMovesContainer
                 game={analyzedGame}
-                setCurrentMove={setCurrentMove}
                 termination={analyzedGame.termination}
-                currentMaiaModel={currentMaiaModel}
               />
             </div>
             <div>
@@ -609,10 +604,9 @@ const Analysis: React.FC<Props> = ({
           <DownloadModelModal progress={maiaProgress} download={downloadMaia} />
         ) : null}
       </AnimatePresence>
-
-      <GameControllerContext.Provider value={{ ...controller }}>
+      <AnalysisGameControllerContext.Provider value={{ ...controller }}>
         {analyzedGame && (isMobile ? mobileLayout : desktopLayout)}
-      </GameControllerContext.Provider>
+      </AnalysisGameControllerContext.Provider>
     </>
   )
 }
